Typo error fixed.
There is a typo error in the second paragraph of the introductory section. The second line of this paragraph is written as, "This is by design - JSX by is not a templating library, it's declarative syntactic sugar over functional JavaScript expressions."
The correct line could be written as, " This is by design - JSX is not a templating library, it's declarative syntactic sugar over functional JavaScript expressions." (avoid using 'by' after JSX).

Thank you.